cli: use logger to print status messages
When cli is run with the output directed to stdout, it may become quite
inconvenient for the progress status messages to pop up on the stdout
as they are being mixed with the audit formatted results.

This change makes both `status` and `statusEnd` events be emitted on
`log.events`, and replaces `console.time` and `console.timeEnd` with
outputting the timers using lighthouse logger, so that all timer
messages go to the same place, usually stderr.

In any event, this change makes it easier to control and redirect
status messages in the same way all other logging is done.