Ensure HTML is written to disk when outputMode=pretty.
`outputMode` in this case is derived from the CLI options and ends up being a regular string.
Then it's being compared to an enum value, which in Typscript is a number by default.

This PR ensures that we compare string against string and correctly write the HTML output to disk.