update helpText for title audit
Since the audit occurs in both the a11y and SEO categories, the new helpText now mentions both. The helpText links to a new reference on d.g.c/web that discusses both the a11y and SEO aspects of the audit. The old aXe doc previously linked to only mentioned the a11y side.