core(tracehouse): Use requestedUrl to detect navigationStart
The TraceProcessor was using a regex to find the first `navigationStart`
event "of interest", excluding navigation events that did not represent
the page under test (e.g. requests from chrome extensions). It now
compares the `documentLoaderURL` of the event directly with the
requestedUrl instead.

This should help prevent false positive `navigationStart` events from
being incorrectly identified as the start of the navigation of interest.
It also prevents a false negative in the case where a chrome-extension
URL is under test, thus fixing #9402