report(drop-down-menu): close menu when focus is lost
Summary
Problem: The report drop down menu does not close when the escape key
is pressed in the devtools.

This creates a challenge for customer relying on the keyboard that wish
to dismiss the drop down.

This also does not align with the experience of the WAI ARIA practice
navigation menu button example [1].

Root cause: The escape key event is captured and not propagated by the
devtools to toggle the console drawer.

Solution: Close the menu when focus moves out of the menu.

This better aligns the menu with the Navigation menu button example for
which its interaction design is based on [1].

Code Change
* Add a menu focus out handler to close the drop down menu
* Unfortunetly JSDOM does not support focusout events so unit tests
  are not possible at this time [2].

[1]: https://www.w3.org/TR/2019/NOTE-wai-aria-practices-1.1-20190207/examples/menu-button/menu-button-links.html
[2]: https://github.com/jsdom/jsdom/issues/2708

Related PR
9433