fix: quantize DPRs
In some contexts lighthouse uses a DPR of 2.625, which is a rather uncommon DPR, so to minimize the number of failures, we will quantize the original DPR to the lowest integer. That strategy is also used for large images. In which effectively we ask for a DPR which is 1 below the original DPR.