Make raven an optional peer dependency
This package has been deprecated in favor of `@sentry/node`. It just hasn’t been
marked as such on npmjs.com.

It depends on deprecated packages, which yields deprecation warnings for users
installing lighthouse.

It’s usage is optional in lighthouse. It’s disabled by default, and even if it’s
enabled in options, the `require` statement for it is in a try-catch block.